Code cleaning on Identity
